Fix test APIs anf a bug with data buffer
