+++ java/src/main/java/ai/rapids/cudf/HostColumnVectorCore.java
@@ -21,9 +21,12 @@ package ai.rapids.cudf;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+
+import static ai.rapids.cudf.HostColumnVector.OFFSET_SIZE;
 
 /**
  * A class that holds Host side Column Vector APIs and the OffHeapState.
@@ -150,6 +153,215 @@ public class HostColumnVectorCore implements AutoCloseable {
     } else {
       return new String();
     }
+  }
+
+
+  /////////////////////////////////////////////////////////////////////////////
+  // DATA ACCESS
+  /////////////////////////////////////////////////////////////////////////////
+
+  /**
+   * For testing only.  Allows null checks to go past the number of rows, but not past the end
+   * of the buffer.  NOTE: If the validity vector was allocated by cudf itself it is not
+   * guaranteed to have the same padding, but for all practical purposes it does.  This is
+   * just to verify that the buffer was allocated and initialized properly.
+   */
+  boolean isNullExtendedRange(long index) {
+    long maxNullRow = BitVectorHelper.getValidityAllocationSizeInBytes(rows) * 8;
+    assert (index >= 0 && index < maxNullRow) : "TEST: index is out of range 0 <= " + index + " <" +
+        " " + maxNullRow;
+    if (hasValidityVector()) {
+      if (nullCount.isPresent() && !hasNulls()) {
+        return false;
+      }
+      return BitVectorHelper.isNull(offHeap.valid, index);
+    }
+    return false;
+  }
+
+  /**
+   * Get access to the raw host buffer for this column.  This is intended to be used with a lot
+   * of caution.  The lifetime of the buffer is tied to the lifetime of the column (Do not close
+   * the buffer, as the column will take care of it).  Do not modify the contents of the buffer or
+   * it might negatively impact what happens on the column.  The data must be on the host for this
+   * to work.
+   * @param type the type of buffer to get access to.
+   * @return the underlying buffer or null if no buffer is associated with it for this column.
+   * Please note that if the column is empty there may be no buffers at all associated with the
+   * column.
+   */
+  public HostMemoryBuffer getHostBufferFor(BufferType type) {
+    HostMemoryBuffer srcBuffer = null;
+    switch(type) {
+      case VALIDITY:
+        srcBuffer = offHeap.valid;
+        break;
+      case OFFSET:
+        srcBuffer = offHeap.offsets;
+        break;
+      case DATA:
+        srcBuffer = offHeap.data;
+        break;
+      default:
+        throw new IllegalArgumentException(type + " is not a supported buffer type.");
+    }
+    return srcBuffer;
+  }
+
+  void copyHostBufferBytes(byte[] dst, int dstOffset, BufferType src, long srcOffset,
+                           int length) {
+    assert dstOffset >= 0;
+    assert srcOffset >= 0;
+    assert length >= 0;
+    assert dstOffset + length <= dst.length;
+
+    HostMemoryBuffer srcBuffer = getHostBufferFor(src);
+
+    assert srcOffset + length <= srcBuffer.length : "would copy off end of buffer "
+        + srcOffset + " + " + length + " > " + srcBuffer.length;
+    UnsafeMemoryAccessor.getBytes(dst, dstOffset,
+        srcBuffer.getAddress() + srcOffset, length);
+  }
+  /**
+   * Generic type independent asserts when getting a value from a single index.
+   * @param index where to get the data from.
+   */
+  private void assertsForGet(long index) {
+    assert (index >= 0 && index < rows) : "index is out of range 0 <= " + index + " < " + rows;
+    assert !isNull(index) : " value at " + index + " is null";
+  }
+
+  /**
+   * Get the value at index.
+   */
+  public byte getByte(long index) {
+    assert type == DType.INT8 || type == DType.UINT8 || type == DType.BOOL8 : type +
+        " is not stored as a byte.";
+    assertsForGet(index);
+    return offHeap.data.getByte(index * type.sizeInBytes);
+  }
+
+  /**
+   * Get the value at index.
+   */
+  public final short getShort(long index) {
+    assert type == DType.INT16 || type == DType.UINT16 : type + " is not stored as a short.";
+    assertsForGet(index);
+    return offHeap.data.getShort(index * type.sizeInBytes);
+  }
+
+  /**
+   * Get the value at index.
+   */
+  public final int getInt(long index) {
+    assert type.isBackedByInt() : type + " is not stored as a int.";
+    assertsForGet(index);
+    return offHeap.data.getInt(index * type.sizeInBytes);
+  }
+
+  /**
+   * Get the starting byte offset for the string at index
+   */
+  long getStartStringOffset(long index) {
+    assert type == DType.STRING: type + " is not a supported string type.";
+    assert (index >= 0 && index < rows) : "index is out of range 0 <= " + index + " < " + rows;
+    return offHeap.offsets.getInt(index * 4);
+  }
+
+  /**
+   * Get the ending byte offset for the string at index.
+   */
+  long getEndStringOffset(long index) {
+    assert type == DType.STRING : type + " is not a supported string type.";
+    assert (index >= 0 && index < rows) : "index is out of range 0 <= " + index + " < " + rows;
+    // The offsets has one more entry than there are rows.
+    return offHeap.offsets.getInt((index + 1) * 4);
+  }
+
+  /**
+   * Get the value at index.
+   */
+  public final long getLong(long index) {
+    // Timestamps with time values are stored as longs
+    assert type.isBackedByLong(): type + " is not stored as a long.";
+    assertsForGet(index);
+    return offHeap.data.getLong(index * type.sizeInBytes);
+  }
+
+  /**
+   * Get the value at index.
+   */
+  public final float getFloat(long index) {
+    assert type == DType.FLOAT32 : type + " is not a supported float type.";
+    assertsForGet(index);
+    return offHeap.data.getFloat(index * type.sizeInBytes);
+  }
+
+  /**
+   * Get the value at index.
+   */
+  public final double getDouble(long index) {
+    assert type == DType.FLOAT64 : type + " is not a supported double type.";
+    assertsForGet(index);
+    return offHeap.data.getDouble(index * type.sizeInBytes);
+  }
+
+  /**
+   * Get the boolean value at index
+   */
+  public final boolean getBoolean(long index) {
+    assert type == DType.BOOL8 : type + " is not a supported boolean type.";
+    assertsForGet(index);
+    return offHeap.data.getBoolean(index * type.sizeInBytes);
+  }
+
+  /**
+   * Get the raw UTF8 bytes at index.  This API is faster than getJavaString, but still not
+   * ideal because it is copying the data onto the heap.
+   */
+  public byte[] getUTF8(long index) {
+    assert type == DType.STRING : type + " is not a supported string type.";
+    assertsForGet(index);
+    int start = offHeap.offsets.getInt(index * OFFSET_SIZE);
+    int size = offHeap.offsets.getInt((index + 1) * OFFSET_SIZE) - start;
+    byte[] rawData = new byte[size];
+    if (size > 0) {
+      offHeap.data.getBytes(rawData, 0, start, size);
+    }
+    return rawData;
+  }
+
+  /**
+   * Get the value at index.  This API is slow as it has to translate the
+   * string representation.  Please use it with caution.
+   */
+  public String getJavaString(long index) {
+    byte[] rawData = getUTF8(index);
+    return new String(rawData, StandardCharsets.UTF_8);
+  }
+
+  /**
+   * WARNING: Strictly for test only. This call is not efficient for production.
+   */
+  List getList(long rowIndex) {
+    assert rowIndex < rows;
+    assert type == DType.LIST;
+    List retList = new ArrayList();
+    int start = offHeap.offsets.getInt(rowIndex * DType.INT32.getSizeInBytes());
+    int end = offHeap.offsets.getInt((rowIndex + 1) * DType.INT32.getSizeInBytes());
+    // check if null or empty
+    if (start == end) {
+      if (isNull(rowIndex)) {
+        return null;
+      }
+    }
+    for(int j = start; j < end; j++) {
+      for (HostColumnVectorCore childHcv : children) {
+        // lists have only 1 child
+        retList.add(childHcv.getElement(j));
+      }
+    }
+    return retList;
   }
 
   /**
